Marketplace server/requirements.txt
# Core
fastapi==0.109.1
uvicorn==0.27.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
asyncpg==0.29.0
alembic==1.13.1
pydantic==2.11.7
pydantic-settings==2.10.1

# Auth
python-jose[cryptography]>=3.3.1  
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# Optional
redis==5.0.1
aioredis==2.0.1
celery==5.3.6
flower==2.0.1

# Docs
mkdocs==1.5.3
mkdocs-material==9.5.3

# Tests
pytest==7.4.4
pytest-asyncio==0.23.5
httpx==0.26.0

# Dotevn
python-dotenv==1.1.1